Move parser to a package of its own
